+++ app/(team)/[id]/todo/_components/TodoItem.tsx
@@ -1,5 +1,6 @@
 import DateTimeFrequency from "@/app/(team)/[id]/todo/_components/DateTimeFrequency";
 import API from "@/app/_api";
+import Icon from "@/app/_icons";
 import { convertIsoToDateAndTime } from "@/app/_utils/IsoToFriendlyDate";
 import Image from "next/image";
 import { useRef } from "react";
@@ -20,7 +21,8 @@ type TodoItemProps = {
 	currentDate: Date;
 	taskId: number;
 	onToggleTodo: (todoId: number, doneAt: string) => void;
-	onClick: (groupId: number, taskId: number, todoId: number, date: Date, doneAt: string) => void;
+	onTodoClick: (groupId: number, taskId: number, todoId: number, date: Date, doneAt: string) => void;
+	onTodoDelete: (todoId: number) => void;
 };
 
 /* eslint-disable jsx-a11y/click-events-have-key-events */
@@ -28,7 +30,7 @@ type TodoItemProps = {
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 
-export default function TodoItem({ taskId, todoItem, groupId, currentDate, onToggleTodo, onClick }: TodoItemProps) {
+export default function TodoItem({ taskId, todoItem, groupId, currentDate, onToggleTodo, onTodoClick, onTodoDelete }: TodoItemProps) {
 	const isLongPress = useRef(false);
 	const timerId = useRef<NodeJS.Timeout>();
 
@@ -43,54 +45,66 @@ export default function TodoItem({ taskId, todoItem, groupId, currentDate, onTog
 	const { date, time } = convertIsoToDateAndTime(todoItem.date); // 날짜 변환
 	return (
 		<div
-			className="flex w-full flex-col gap-[11px] rounded-lg bg-background-secondary px-[14px] py-3 hover:cursor-pointer hover:bg-background-tertiary"
+			className="flex w-full justify-between rounded-lg bg-background-secondary px-[14px] py-3 hover:cursor-pointer hover:bg-background-tertiary"
 			key={todoItem.id}
 			onClick={(e) => {
 				e.stopPropagation();
 				if (isLongPress.current) return;
-				onClick(groupId, taskId, todoItem.id, currentDate, todoItem.doneAt);
+				onTodoClick(groupId, taskId, todoItem.id, currentDate, todoItem.doneAt);
 			}}
 			onMouseDown={handleMouseDown}
 			onMouseUp={() => {
 				clearTimeout(timerId.current);
 			}}
 		>
-			<div className="flex items-center justify-between">
-				<div className="flex gap-3">
-					<button
-						className="group relative"
-						type="button"
-						aria-label="todo-done"
-						onClick={(event) => {
-							event.stopPropagation();
-							onToggleTodo(todoItem.id, todoItem.doneAt);
-						}}
-					>
-						<Image src={todoItem.doneAt ? "/icons/checkBox.svg" : "/icons/uncheckBox.svg"} alt="Todo status" width={24} height={24} />
+			<div className="flex flex-col gap-[11px]">
+				<div className="flex items-center justify-between">
+					<div className="flex gap-3">
+						<button
+							className="group relative"
+							type="button"
+							aria-label="todo-done"
+							onClick={(event) => {
+								event.stopPropagation();
+								onToggleTodo(todoItem.id, todoItem.doneAt);
+							}}
+						>
+							<Image src={todoItem.doneAt ? "/icons/checkBox.svg" : "/icons/uncheckBox.svg"} alt="Todo status" width={24} height={24} />
 
-						{/* 호버 시 나타나는 오버레이 */}
-						<div className="absolute inset-0 flex items-center justify-center opacity-0 transition-opacity duration-200 ease-in-out group-hover:opacity-90">
-							<Image src="/icons/checkBox.svg" alt="Preview" width={24} height={24} className="opacity-100" />
-						</div>
-					</button>
-					<div className="group relative inline-block">
-						<button type="button" className={`${todoItem.doneAt ? "line-through" : ""} text-text-primary`}>
-							{todoItem.name}
+							{/* 호버 시 나타나는 오버레이 */}
+							<div className="absolute inset-0 flex items-center justify-center opacity-0 transition-opacity duration-200 ease-in-out group-hover:opacity-90">
+								<Image src="/icons/checkBox.svg" alt="Preview" width={24} height={24} className="opacity-100" />
+							</div>
 						</button>
+						<div className="group relative inline-block">
+							<button type="button" className={`${todoItem.doneAt ? "line-through" : ""} text-text-primary`}>
+								{todoItem.name}
+							</button>
 
-						{/* 툴팁 */}
-						<div className="pointer-events-none absolute bottom-3/4 left-10 mb-2 w-max -translate-x-1/2 transform rounded bg-background-secondary p-2 text-sm text-white opacity-0 transition-opacity duration-200 ease-in-out group-hover:opacity-100">
-							할일 상세 보기
+							{/* 툴팁 */}
+							<div className="pointer-events-none absolute bottom-3/4 left-10 mb-2 w-max -translate-x-1/2 transform rounded bg-background-secondary p-2 text-sm text-white opacity-0 transition-opacity duration-200 ease-in-out group-hover:opacity-100">
+								할일 상세 보기
+							</div>
 						</div>
-					</div>
 
-					<div className="flex items-center justify-center gap-1 text-xs font-normal text-text-default">
-						<Image src="/icons/comment.svg" alt="comment" width={16} height={16} />
-						{todoItem.commentCount}
+						<div className="flex items-center justify-center gap-1 text-xs font-normal text-text-default">
+							<Image src="/icons/comment.svg" alt="comment" width={16} height={16} />
+							{todoItem.commentCount}
+						</div>
 					</div>
 				</div>
+				<DateTimeFrequency date={date} time={time} frequency={frequency[todoItem.frequency]} />
 			</div>
-			<DateTimeFrequency date={date} time={time} frequency={frequency[todoItem.frequency]} />
+			<button
+				aria-label="할 일 삭제"
+				type="submit"
+				onClick={(event) => {
+					event.stopPropagation();
+					onTodoDelete(todoItem.id);
+				}}
+			>
+				<Icon.TodoDelete width={24} height={24} />
+			</button>
 		</div>
 	);
 }
